Battle state now works (kind of)
Press 'Space' to initialize battle. If you run into one on your own, the
character will stop moving, but when I make the battles totally separate
from the player it will work. Getting into a battle will be handled by
the World, and the player object is going to only be for the maps.