We're back baby. Functionality is now back to before I broke it
All states now work correctly. Still untying things, but it works right
now and is easier to change. Changed the MouseManager to let us check if
a mouse button is down or just clicked once. Will add comments next
commit. It also needed to be added as a mouselistener AND a
mouseMotionListener to the canvas. Before with the mouse manager, it
would detect a click and then set the bool to true, but the bool was
never set to false. So it was ALWAYS reporting that it was true. Plus,
the mouseMove event wasn't firing because it wasn't registered as a
mouseMotionListener on the canvas. So when you clicked, it would
constantly be saying that it was clicking but it was still reporting the
location of the last click, even after you moved the mouse. Phew.