No random battles - press space right now
Random battle isn't triggered, but you can press space to start one.
Eventually the gamestate (should be renamed overworld state) will be the
one controlling when a battle is entered. One way to do that will be for
it to keep track of how long the player is walking and then if it hits a
certain threshold (say, 3 seconds) then it enters a battle. We can play
with it though. Sorry random battles are broken right now :(