Added Dialog State, tons of cleanup, and added JavaDocs!
A new DialogState class has been created and added to the StateManager
in order to display dialog. Currently, pressing "Enter" will swtich to
the DialogState, which now displays a filled-in rectangle with sample
text. Pressing Enter again will go back to the GameState. In order to
keep what is currently rendered and just have the dialog on top, we will
need to change the render methods in several classes to not clear the
canvas. I also went through and removed a bunch of unnecessary variables
and objects no longer needed, or some that were never even needed. I
have also gone through and changed all comments explaining
classes/methods to Javadoc comments, as well as adding a Javadoc comment
to each class with a basic description of what each one does. Now, if we
decide to share this project with anyone else, we can now generate a
Javadoc to help them catch up to speed on what everything does (In
Eclipse, Project->Generate Javadoc). I made a bunch of other minor
changes, but I do not remember them all... so many changes...