Dialog now shows over map, dialog now shows gradually.
The guy is still spinning...anyway...now the dialog will show gradually
(we can change the speed or make it variable speed, easy to do. We need
a config file/class eventually anyway). If you press talk, it will all
appear. If the entire message is showing, pressing talk will exit the
state.

Also, trying to play a track that is already playing will not do
anything. I'll add a resetTrack option eventually if we need it.