Updated TODO, fixed dialog display
Before, if the last word of a sentence would push the sentence off the
screen, it wouldn't wrap to the next line. Now it does! Also, I don't
know why it says I edited world3. I put it back exactly how it was.
Maybe it's a tab vs spaces thing.