Updated BattleState to allow spells and added "the finger"
1) BattleState- Added a new "spell" option to cast spells, which right
now is just attacking using the Intelligence stat. This is just a base,
and will eventually have a sub-menu with different spells to cast (that
actually use MP when we rework that whole system). I also added "the
finger" from the TitleState to this state along with the sound effects
and filling the empty parts of the canvas with white.
2) TitleState - removed unused imports
3) Player - changed base intelligence to 20 (to cheese battles for
testing)
4) Updated TODO