Added dialog upon entering the BattleState
Added dialog upon entering the BattleState. For now, I have added it to
the update method, as adding it on the onEnter method displays it before
the BattleState is first rendered. Eventually, this should be made in a
way that the Dialog can somehow be passed an Entity name(s) or
object(s).