Dialogue now roughly works for hitting Talk next to NPCs.
The only issue that I see in my 10 seconds of testing is that if you are
moving and press talk, then when you exit the dialogue state, the camera
will have to recenter on the player. I think it's just the camera
catching up, so I'll try to fix it tonight.