Added MenuState, updated TitleState logic, added a new method to Utilities
1) MenuState - added a fully working menu. This definitely needs to look
prettier, but I am not so great at that.
2) TitleState - Fixed an issue where the highlightLengthMax variable was
only looking at the first option for the maximum string length
(resulting in words longer than the first choice not being highlighted).
This has been changed to use a new method "getLongestString" in
Utilities.
3) Added the getLongestString method in Utilities. This is passed a
String Array and returns the longest string in the array.
4) Map - Removed unused imports.